fix: :bug: исправить reference-дочки в sync конфигурации
